requirements.txt
pandas==0.16.0
python-geohash==0.8.5
scikit-learn==0.18